Renamed connect file and implemented basefunctions for databasemanagement
The db_connect.go file is now named db_info.go to more accurately describe the contents. The functions for GetAll, DropDatabase and CreateCollection have also been implemented.
